feat: useCellSelectionV2フックを抽出
EditableTableV2からセル選択ロジックを独立したカスタムフックに抽出。
これによりコードの再利用性とテストの容易性が向上。

変更内容:
- useCellSelectionV2.ts: セル選択、一括編集、コピー&ペースト機能を提供
- useCellSelectionV2.spec.ts: 10個のユニットテストを追加
- EditableTableV2.tsx: 172行のセル選択ロジックを削除し、フックを使用
- types.spec.ts: 未使用のRowIndexCellPropsインポートを削除

テスト結果:
- ユニットテスト: 10/10 パス
- AppV2 Storybookテスト: 全テストパス
- 型チェック: エラーなし
- リント: エラーなし

Phase 2-1完了
